refactor: move gateway auto discover to a new reconciler
This commit fixes an issue with bgp gateway auto discovery.
Previously there was an issue with the priority of the reconcilers
where the gateway ip set by the neighbor reconciler auto discovery
is not picked up by the pod cidr reconciler when setting the
policy.
This fix refactors the auto discovery logic to a new reconciler that
runs first.

Fixes: #41904
